Refactor by grouping collectors (#6944)
