+++ ibm_mq/datadog_checks/ibm_mq/collectors/__init__.py
@@ -1,0 +1,7 @@
+# (C) Datadog, Inc. 2018-present
+# All rights reserved
+# Licensed under a 3-clause BSD style license (see LICENSE)
+from .channel_metric_collector import ChannelMetricCollector
+from .queue_metric_collector import QueueMetricCollector
+
+__all__ = ['ChannelMetricCollector', 'QueueMetricCollector']
