+++ ibm_mq/datadog_checks/ibm_mq/collectors/channel_metric_collector.py
@@ -6,7 +6,7 @@ from six import iteritems
 
 from datadog_checks.base import AgentCheck, ensure_bytes, ensure_unicode
 
-from . import metrics
+from .. import metrics
 
 try:
     import pymqi
