+++ ibm_mq/datadog_checks/ibm_mq/ibm_mq.py
@@ -8,9 +8,8 @@ from datadog_checks.base import AgentCheck
 from datadog_checks.ibm_mq.metrics import COUNT, GAUGE
 
 from . import connection, errors
-from .channel_metric_collector import ChannelMetricCollector
+from .collectors import ChannelMetricCollector, QueueMetricCollector
 from .config import IBMMQConfig
-from .queue_metric_collector import QueueMetricCollector
 
 try:
     import pymqi
